IT WORKS

+++ particle-sim/inc/particleSystem.h
@@ -13,11 +13,11 @@ public:
 
 	virtual ~ParticleSystem() {}
 
-	virtual double* getPositions() = 0;
+	virtual float* getPositions() = 0;
 
 	virtual unsigned int* getColors() = 0;
 
-	virtual void update(double timeDelta) = 0;
+	virtual void update(float timeDelta) = 0;
 
 	virtual void writecurpostofile(char* file) = 0;
 	
@@ -28,8 +28,8 @@ protected:
 	int p_numParticles;
 
 	// Particle Data
-	double* positions; // 1D Array containing spacial data of each particle (positionElementsCount * numParticles)
-	double* velocities; // 1D Array containing velocity data of each particle (velocityElementsCount * numParticles)
+	float* positions; // 1D Array containing spacial data of each particle (positionElementsCount * numParticles)
+	float* velocities; // 1D Array containing velocity data of each particle (velocityElementsCount * numParticles)
 	unsigned int* colors; // 1D Array containing RGB data of each particle (colorElementsCount * numParticles)
 	unsigned char* particleType; // 1D Array which denotes particle type (0 = Electron; 1 = Proton, 2 = Neutron)
